# PERF: `to_set` speeds up especial for large data