<!--
Thanks for contributing a pull request!

Please follow these standard acronyms to start the commit message:

- ENH: enhancement
- BUG: bug fix
- DOC: documentation
- TYP: type annotations
- TST: addition or modification of tests
- MAINT: maintenance commit (refactoring, typos, etc.)
- BLD: change related to building
- REL: related to releasing
- API: an (incompatible) API change
- DEP: deprecate something, or remove a deprecated object
- DEV: development tool or utility
- REV: revert an earlier commit
- PERF: performance improvement
- BOT: always commit via a bot
- CI: related to CI or CD
- CLN: Code cleanup
-->

- [ ] closes #xxxx
- [x] whatsnew entry


| data                 | `set(s)`         | `set(s.unique())` |
| -------------------- | ---------------- | ----------------- |
| small, `[1, 1]`      | 1.94 µs ± 158 ns | 30.1 µs ± 270 ns  |
| large, `[1, 1]*1000` | 172 µs ± 29.5 µs | 46.8 µs ± 816 ns  |